mmc-utils: Add Read-the-Docs configuration
Prior of building the documentation page, read-the-docs expects a github
project to be imported into it, so I had to mirror the mmc-utils repo
into https://github.com/avri-altman-wdc/mmc-utils.  It also expects some
mandatory settings in a configuration file - .readthedocs.yaml, add
those as well.  Finally, the public documentation is in
https://mmc-utils.readthedocs.io/en/latest/

Signed-off-by: Avri Altman <avri.altman@wdc.com>
Link: https://lore.kernel.org/r/20240624112542.25657-3-avri.altman@wdc.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>