mmc-utils: lsmmc: Pass program name to usage function
This commit updates the usage function to accept the program name as an
argument, allowing the usage message to display the correct program name
dynamically. This change improves the clarity and accuracy of the usage
message. This enhancement ensures that the usage message accurately
reflects the name of the register: cid, csd or scr, being read.

Signed-off-by: Avri Altman <avri.altman@sandisk.com>
Link: https://lore.kernel.org/r/20250310055231.304728-4-avri.altman@sandisk.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>