mmc-utils: lsmmc: Update docs to include new options
This commit updates the man and HOWTO page entries to include the new
options for the csd, cid, and scr commands. The updated entries reflect
the changes allowing the use of bus_type and register content as
arguments. These updates provide users with clear instructions on how to
use the new options for the csd, cid, and scr read commands.

Signed-off-by: Avri Altman <avri.altman@sandisk.com>
Link: https://lore.kernel.org/r/20250310055231.304728-8-avri.altman@sandisk.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>